% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_methods.R
\name{compare}
\alias{compare}
\alias{compare.growmod}
\alias{compare.growmod_cv}
\alias{compare.growmod_multi}
\alias{compare.growmod_cv_multi}
\title{Comparison method for \code{"growmod"} objects}
\usage{
compare(x, ...)

\method{compare}{growmod}(..., x)

\method{compare}{growmod_cv}(..., x)

\method{compare}{growmod_multi}(..., x)

\method{compare}{growmod_cv_multi}(..., x)
}
\arguments{
\item{x}{an object of class \code{"growmod"}, \code{"growmod_cv"},
\code{"growmod_multi"} or \code{"growmod_cv_multi}}

\item{\dots}{additional objects of class \code{"growmod"}, \code{"growmod_cv"},
\code{"growmod_multi"} or \code{"growmod_cv_multi}}
}
\description{
Compare multiple fitted growmod models
}
\section{Methods (by class)}{
\itemize{
\item \code{growmod}: compare fitted \code{"growmod"} models

\item \code{growmod_cv}: compare fitted \code{"growmod_cv"} models

\item \code{growmod_multi}: compare fitted \code{"growmod_multi"} models

\item \code{growmod_cv_multi}: compare fitted \code{"growmod_cv_multi"} models
}}

\examples{
\dontrun{

# simulate some data
data_test <- growmod_sim()

# fit a growmod model
mod1 <- growmod(size ~ (index | block / predictors),
                data = data_sim,
                model = 'hillslope',
                n_iter = 100,
                n_burnin = 50,
                n_chains = 1,
                stan_cores = 1)

# fit a second model
mod2 <- growmod(size ~ (index | block / predictors),
                data = data_sim,
                model = 'koblog',
                n_iter = 100,
                n_burnin = 50,
                n_chains = 1,
                stan_cores = 1)

# compare the two fitted models
compare(mod1, mod2)
}

}
