% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_fit.R
\name{growmod-fit}
\alias{growmod-fit}
\alias{growmod}
\alias{growmod.formula}
\alias{growmod.default}
\title{Fitting growth models using growmod}
\usage{
growmod(x, ...)

\method{growmod}{formula}(x, data = NULL, model = "hillslope",
  n_iter = 5000, n_burnin = floor(n_iter/2), n_thin = 1,
  n_chains = 4, stan_cores = 1, spline_params = list(degree = 8,
  n_knots = 10, spline_type = "ispline"), ...)

\method{growmod}{default}(x, index, block, predictors, model, n_iter,
  n_burnin, n_thin, n_chains, stan_cores, spline_params, call = NULL,
  formula = NULL, ...)
}
\arguments{
\item{x}{model formula (see details) or vector of size data}

\item{\dots}{parameters to be passed to stan model call}

\item{data}{list containing data on size, age, species, and traits}

\item{model}{growth model form to be fitted}

\item{n_iter}{number of HMC iterations to run for stan model}

\item{n_burnin}{number of HMC iterations to run for stan model}

\item{n_thin}{thinning rate for HMC chains}

\item{n_chains}{number of HMC chains}

\item{stan_cores}{number of local cores to use in stan model fitting}

\item{spline_params}{named list of settings for spline model (degree, n_knots, spline_type)}

\item{index}{list containing data on size, age, species, and traits}

\item{block}{list containing data on size, age, species, and traits}

\item{predictors}{list containing data on size, age, species, and traits}

\item{call}{model call passed from \code{growmod.formula}}

\item{formula}{model formula passed from \code{growmod.formula}}
}
\value{
\code{mod} A fitted \code{growmod} object containing
  parameter estimates, validation statistics, and the original
  data set used to fit the model.
  \item{fitted}{fitted size values}
  \item{r2}{r-squared based on Pearson's r between observed and fitted size data}
  \item{rmsd}{root-mean-squared deviance of the fitted model}
  \item{md}{mean deviance of the fitted model}
  \item{loo}{leave-one-out information criterion (deviance-based approximation to leave-one-out cross validation)}
  \item{waic}{Watanabe-Akaike (widely applicable) information criterion}
  \item{stan_summary}{summary of HMC samples for each parameter}
  \item{data_set}{full data set used in model fitting}
  \item{predictors}{predictor variables used in model fitting}
  \item{model}{growth model used in model fitting}
  \item{stanmod}{written Stan model}
  \item{spline_params}{parameters for spline model}
  \item{stan_cores}{number of cores used by Stan}
  \item{n_iter}{number of HMC iterations (burnin + sampling)}
  \item{n_burnin}{number of HMC iterations to discard before storing samples}
  \item{n_thin}{thinning rate for HMC chains (store every n_thin iterations)}
  \item{n_chains}{number of HMC chains}
  \item{formula}{model formula}
  \item{call}{original model call}
}
\description{
Fit a growth model to data on sizes through time. Currently
  uses Stan to fit models for one of twelve common growth curves.
}
\details{
\code{growmod} takes a formula or data vectors as arguments
  and constructs and fits a Stan model for a chosen growth curve. The formula
  syntax is standard across all growmod models. A growmod formula specifies the
  size variable as a response and an index variable (e.g. age) as a predictor.
  A growmod formula can specify blocks in the data; the syntax for this is as
  if setting a random effect for the index. If specified, a separate growth curve
  will be fitted for each block. In addition, a growmod formula can include
  predictor variables within blocks, which allows prediction of growth curves
  to new blocks (see examples, below).
}
\examples{
\dontrun{
  # simulate some data
  data_sim <- growmod_sim(n = 100,
                          nblock = 5,
                          age_range = c(0, 50),
                          include_predictors = TRUE,
                          true_model = 'hillslope')

  # fit the correct model
  mod1 <- growmod(size ~ (index | block / predictors),
                  data = data_sim,
                  model = 'hillslope',
                  n_iter = 100,
                  n_burnin = 50,
                  n_chains = 1,
                  stan_cores = 1)

  # plot the fitted model
  plot(mod1)
  
  # summarise the fitted model
  summary(mod1)
  
  # print the fitted model
  print(mod1)

  # cross validate the fitted model
  mod1_cv <- validate(mod1, n_cv = 'loo')
                   
  # fit an incorrect model
  mod2 <- growmod(size ~ (index | block / predictors),
                  data = data_sim,
                  model = 'koblog',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # compare the fitted models using summary measures of fit
  compare(mod1, mod2)
  
  # compare the fitted models using cross validation
  mod2_cv <- validate(mod2)
  compare(mod1_cv, mod2_cv)
  
  # example of different predictor sets for each parameter
  data_sim2 <- with(data_sim, list(block = block,
                                   size = size,
  mod3 <- growmod(size ~ (index | block / predictors),
                  data = data_sim2,
                  model = 'koblog',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)
  
  # example of multiple models fitted in one call
  mod_multi <- growmod(size ~ (index | block / predictors),
                       data = data_sim,
                       model = c('hillslope',
                                 'koblog'),
                       n_iter = 1000,
                       n_burnin = 500,
                       n_chains = 2,
                       stan_cores = 1)
}

}
