% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{growmod-package}
\alias{growmod}
\alias{growmod-package}
\title{Fit and validate growth models for a suite of common growth curves}
\description{
growmod allows you to fit growth models to size-at-age data and to 
  validate these models with cross validation or common model
  comparison tools (r2, WAIC, loo).
  
The main function you'll use in \pkg{growmod} is \code{\link{growmod}}
  and its associated methods. See the package vignettes and help
  files for examples of fitting growth curves to data.
}
\references{
Thomas et al. (in prep.) PAPER.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jdyen/growmod}
  \item Report bugs at \url{https://github.com/jdyen/growmod/issues}
}

}
\author{
\strong{Maintainer}: Jian Yen \email{jdl.yen@gmail.com}

Authors:
\itemize{
  \item Freya Thomas
}

}
