% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_validate.R
\name{growmod-validate}
\alias{growmod-validate}
\alias{validate}
\alias{validate.formula}
\alias{validate.growmod}
\alias{validate.growmod_multi}
\title{Validate growth models fitted with growmod}
\usage{
validate(x, ...)

\method{validate}{formula}(formula, data = NULL, model = "hillslope",
  n_cv = NULL, test_data = NULL, n_iter = 5000,
  n_burnin = floor(n_iter/2), n_thin = 1, n_chains = 4, stan_cores = 1,
  spline_params = list(degree = 8, n_knots = 10, spline_type = "ispline"),
  ...)

\method{validate}{growmod}(x, n_cv = NULL, test_data = NULL,
  n_iter = NULL, n_burnin = NULL, n_thin = NULL, n_chains = NULL, ...)

\method{validate}{growmod_multi}(x, n_cv, train_data = NULL,
  test_data = NULL, n_iter = NULL, n_burnin = NULL, n_thin = NULL,
  n_chains = NULL, ...)
}
\description{
Validate a fitted growth model using cross validation
    or by specifying your own test data set
}
