% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_fit.R, R/growmod_methods.R,
%   R/growmod_package.R, R/growmod_validate.R
\docType{package}
\name{growmod}
\alias{growmod}
\alias{growmod}
\alias{growmod.formula}
\alias{growmod.default}
\alias{compare}
\alias{compare.growmod}
\alias{compare.growmod_cv}
\alias{compare.growmod_multi}
\alias{compare.growmod_cv_multi}
\alias{plot.growmod}
\alias{plot.growmod_cv}
\alias{plot.growmod_multi}
\alias{plot.growmod_cv_multi}
\alias{summary.growmod}
\alias{summary.growmod_cv}
\alias{summary.growmod_multi}
\alias{summary.growmod_cv_multi}
\alias{print.growmod}
\alias{print.growmod_cv}
\alias{print.growmod_multi}
\alias{print.growmod_cv_multi}
\alias{predict.growmod}
\alias{growmod}
\alias{growmod-package}
\alias{validate}
\alias{validate.formula}
\alias{validate.growmod}
\alias{validate.growmod_multi}
\title{Fitting growth models using growmod}
\usage{
growmod(x, ...)

\method{growmod}{formula}(formula, data = NULL, model = "hillslope",
  n_iter = 5000, n_burnin = floor(n_iter/2), n_thin = 1, n_chains = 4,
  stan_cores = 1, spline_params = list(degree = 8, n_knots = 10, spline_type
  = "ispline"), ...)

\method{growmod}{default}(size, index, block, predictors, model, n_iter,
  n_burnin, n_thin, n_chains, stan_cores, spline_params, ...)

compare(x, ...)

\method{compare}{growmod}(..., x)

\method{compare}{growmod_cv}(..., x)

\method{compare}{growmod_multi}(..., x)

\method{compare}{growmod_cv_multi}(..., x)

plot.growmod(x, ...)

plot.growmod_cv(x, ...)

plot.growmod_multi(x, group_blocks = TRUE, ...)

plot.growmod_cv_multi(x, ...)

\method{summary}{growmod}(x, ...)

\method{summary}{growmod_cv}(x, ...)

\method{summary}{growmod_multi}(x, ...)

\method{summary}{growmod_cv_multi}(x, ...)

\method{print}{growmod}(x, ...)

\method{print}{growmod_cv}(x, ...)

\method{print}{growmod_multi}(x, ...)

\method{print}{growmod_cv_multi}(x, ...)

\method{predict}{growmod}(object, newdata, ...)

validate(x, ...)

\method{validate}{formula}(formula, data = NULL, model = "hillslope",
  n_cv = NULL, test_data = NULL, n_iter = 5000,
  n_burnin = floor(n_iter/2), n_thin = 1, n_chains = 4, stan_cores = 1,
  spline_params = list(degree = 8, n_knots = 10, spline_type = "ispline"),
  ...)

\method{validate}{growmod}(x, n_cv = NULL, test_data = NULL,
  n_iter = NULL, n_burnin = NULL, n_thin = NULL, n_chains = NULL, ...)

\method{validate}{growmod_multi}(x, n_cv, train_data = NULL,
  test_data = NULL, n_iter = NULL, n_burnin = NULL, n_thin = NULL,
  n_chains = NULL, ...)
}
\arguments{
\item{\dots}{parameters to be passed to stan model call}

\item{formula}{list containing data on size, age, species, and traits}

\item{data}{list containing data on size, age, species, and traits}

\item{model}{growth model form to be fitted}

\item{n_iter}{number of HMC iterations to run for stan model}

\item{n_burnin}{number of HMC iterations to run for stan model}

\item{n_thin}{thinning rate for HMC chains}

\item{n_chains}{number of HMC chains}

\item{stan_cores}{number of local cores to use in stan model fitting}

\item{spline_params}{named list of settings for spline model (degree, n_knots, spline_type)}

\item{size}{list containing data on size, age, species, and traits}

\item{index}{list containing data on size, age, species, and traits}

\item{block}{list containing data on size, age, species, and traits}

\item{predictors}{list containing data on size, age, species, and traits}
}
\value{
\code{mod} A fitted \code{growmod} object containing
  parameter estimates, validation statistics, and the original
  data set used to fit the model.
}
\description{
Fit a growth model to data on sizes through time. Currently
  uses Stan to fit models for one of twelve common growth curves.

R package to fit growth models to data on size at age. Supports
       multiple model types, including standard growth-curve fitting, growth
       curves that differ among blocks in a data set (e.g. different growth
       curves for each species in a data set), and hierarchical models of growth
       curves where curve parameters depend on a predictor variable (e.g., a
       species's traits). Includes methods for model validation, comparison,
       and plotting.
}
\details{
\code{growmod} takes a formula or data vectors as arguments
  and constructs and fits a Stan model for a chosen growth curve.
}
\examples{
\dontrun {
  # simulate some data
  data_sim <- growth_data_sim(n = 100,
                              nblock = 5,
                              age_range = c(0, 50),
                              include_predictors = TRUE,
                              true_model = 'hillslope')

  # fit the correct model
  mod1 <- growmod(size ~ (age | block / predictors),
                  data = data_sim,
                  model = 'hillslope',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # plot the fitted model
  plot(mod1)
  
  # summarise the fitted model
  summary(mod1)
  
  # print the fitted model
  print(mod1)

  # cross validate the fitted model
  mod1_cv <- validate(mod1)
                   
  # fit an incorrect model
  mod2 <- growmod(size ~ (age | block / predictors),
                  data = data_sim,
                  model = 'koblog',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # compare the fitted models using summary measures of fit
  compare(mod1, mod2)
  
  # compare the fitted models using cross validation
  mod2_cv <- validate(mod2)
  compare(mod1_cv, mod2_cv)
  
  # example of different predictor sets for each parameter
  # TO BE ADDED
  
  # example of multiple models fitted in one call
  # TO BE ADDED
  
  # example of multiple models fitted in one call with different
  #  predictors for each model and parameter
  # TO BE ADDED
  
}

\dontrun {
  # simulate some data
  data_sim <- growth_data_sim(n = 100,
                              nblock = 5,
                              age_range = c(0, 50),
                              include_predictors = TRUE,
                              true_model = 'hillslope')

  # fit the correct model
  mod1 <- growmod(size ~ (age | block / predictors),
                  data = data_sim,
                  model = 'hillslope',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # plot the fitted model
  plot(mod1)
  
  # summarise the fitted model
  summary(mod1)
  
  # print the fitted model
  print(mod1)

  # cross validate the fitted model
  mod1_cv <- validate(mod1)
                   
  # fit an incorrect model
  mod2 <- growmod(size ~ (age | block / predictors),
                  data = data_sim,
                  model = 'koblog',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # compare the fitted models using summary measures of fit
  compare(mod1, mod2)
  
  # compare the fitted models using cross validation
  mod2_cv <- validate(mod2)
  compare(mod1_cv, mod2_cv)
  
  # example of different predictor sets for each parameter
  # TO BE ADDED
  
  # example of multiple models fitted in one call
  # TO BE ADDED
  
  # example of multiple models fitted in one call with different
  #  predictors for each model and parameter
  # TO BE ADDED
  
}
}
\author{
Jian Yen, \email{jdl.yen@gmail.com}

Freya Thomas
}
