% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_sim.R
\name{growmod_sim}
\alias{growmod_sim}
\title{Simulate growth curve data from a chosen growth curve}
\usage{
growmod_sim(n = 100, nblock = 5, age_range = c(0, 20),
  include_predictors = TRUE, true_model = "hillslope")
}
\arguments{
\item{n}{number of simulated growth curves}

\item{nblock}{number of blocks (e.g. species) to include in simulated data}

\item{age_range}{range of ages for simulated data}

\item{include_predictors}{logical; should growth curve parameters depend on predictor
variables that are specific to each block?}

\item{true_model}{the true growth curve model to use when simulating growth curves}
}
\value{
list containing simulated data
  \item{size}{simulated size data}
  \item{block}{block identifier}
  \item{index}{index identifier (e.g. age)}
  \item{predictors}{predictor variables if \code{include_predictors = TRUE}}
  \item{coefs}{true coefficients used for simulated data}
}
\description{
Simulate growth curve data for a given growth curve and
  optionally include blocks within the data set (e.g. species) and 
  predictor varibles that determine the growth parameters for each
  block.
}
\details{
Simulate growth curve data from a chosen growth curve model. This function
  can simulate data from multiple separate blocks (e.g. growth curves for different
  species) and optionally can include predictor variables (e.g. traits) that
  determine the growth parameters for a given block.
}
\examples{
sim_data <- growmod_sim(n = 100,
                        nblock = 5,
                        age_range = c(0, 50),
                        include_predictors = TRUE,
                        true_model = 'hillslope')

}
