% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_methods.R
\name{growmod-methods}
\alias{growmod-methods}
\alias{compare}
\title{Methods for fitted growth models}
\usage{
compare(x, ...)
}
\arguments{
\item{x}{fitted growmod, growmod_cv, growmod_multi, or growmod_cv_multi object}

\item{\dots}{additional fitted growmod objects}
}
\description{
Methods to print, summarise, compare, plot, and extract
  details from fitted growmod models.
}
\details{

}
\section{Usage}{
 \preformatted{

compare(x, ...)
plot(x)
summary(x)
print(x)
fitted(x)
residuals(x)

}
}

\examples{

\dontrun {

}

}
