% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_methods.R
\name{growmod-methods}
\alias{growmod-methods}
\alias{compare}
\alias{compare.growmod}
\alias{compare.growmod_cv}
\alias{compare.growmod_multi}
\alias{compare.growmod_cv_multi}
\alias{plot.growmod}
\alias{plot.growmod_cv}
\alias{plot.growmod_multi}
\alias{plot.growmod_cv_multi}
\alias{summary.growmod}
\alias{summary.growmod_cv}
\alias{summary.growmod_multi}
\alias{summary.growmod_cv_multi}
\alias{print.growmod}
\alias{print.growmod_cv}
\alias{print.growmod_multi}
\alias{print.growmod_cv_multi}
\alias{predict.growmod}
\title{Methods for fitted growth models}
\usage{
compare(x, ...)

\method{compare}{growmod}(..., x)

\method{compare}{growmod_cv}(..., x)

\method{compare}{growmod_multi}(..., x)

\method{compare}{growmod_cv_multi}(..., x)

plot.growmod(x, ...)

plot.growmod_cv(x, ...)

plot.growmod_multi(x, group_blocks = TRUE, ...)

plot.growmod_cv_multi(x, ...)

\method{summary}{growmod}(x, ...)

\method{summary}{growmod_cv}(x, ...)

\method{summary}{growmod_multi}(x, ...)

\method{summary}{growmod_cv_multi}(x, ...)

\method{print}{growmod}(x, ...)

\method{print}{growmod_cv}(x, ...)

\method{print}{growmod_multi}(x, ...)

\method{print}{growmod_cv_multi}(x, ...)

\method{predict}{growmod}(object, newdata, ...)
}
\description{
Compare, plot, summarise, and print models fitted using
  the growmod R package
}
