% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_package.R, R/pkgname.R
\docType{package}
\name{growmod-package}
\alias{growmod-package}
\alias{growmod}
\alias{growmod-package}
\title{growmod: Growth Curve Modelling and Validation with Stan}
\description{
R package to fit growth models to data on size at age. Supports
       multiple model types, including standard growth-curve fitting, growth
       curves that differ among blocks in a data set (e.g. different growth
       curves for each species in a data set), and hierarchical models of growth
       curves where curve parameters depend on a predictor variable (e.g., a
       species's traits). Includes methods for model validation, comparison,
       and plotting.

growmod allows you to fit growth models to size-at-age data and to 
  validate these models with cross validation or common model
  comparison tools (r2, WAIC, loo).
  
The main function you'll use in \pkg{growmod} is \code{\link{growmod}}
  and its associated methods, particularly the \code{\link{validate}} method.
  See the package vignettes and help files for examples of fitting growth
  curves to data.
}
\examples{
\dontrun{
  # simulate some data
  data_sim <- growth_data_sim(n = 100,
                              nblock = 5,
                              age_range = c(0, 50),
                              include_predictors = TRUE,
                              true_model = 'hillslope')

  # fit the correct model
  mod1 <- growmod(size ~ (age | block / predictors),
                  data = data_sim,
                  model = 'hillslope',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # plot the fitted model
  plot(mod1)
  
  # summarise the fitted model
  summary(mod1)
  
  # print the fitted model
  print(mod1)

  # cross validate the fitted model
  mod1_cv <- validate(mod1)
                   
  # fit an incorrect model
  mod2 <- growmod(size ~ (age | block / predictors),
                  data = data_sim,
                  model = 'koblog',
                  n_iter = 1000,
                  n_burnin = 500,
                  n_chains = 2,
                  stan_cores = 1)

  # compare the fitted models using summary measures of fit
  compare(mod1, mod2)
  
  # compare the fitted models using cross validation
  mod2_cv <- validate(mod2)
  compare(mod1_cv, mod2_cv)
  
  # example of different predictor sets for each parameter
  # TO BE ADDED
  
  # example of multiple models fitted in one call
  mod_multi <- growmod(size ~ (age | block / predictors),
                       data = data_sim,
                       model = c('hillslope',
                                 'koblog'),
                       n_iter = 1000,
                       n_burnin = 500,
                       n_chains = 2,
                       stan_cores = 1)
  
  # example of multiple models fitted in one call with different
  #  predictors for each model and parameter
  # TO BE ADDED
}
}
\references{
Thomas, F.M., Yen, J.D.L. and Vesk, P.A. (in prep.) Which nonlinear model?
  Using cross-validation to evaluate growth models in ecology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jdyen/growmod}
  \item Report bugs at \url{https://github.com/jdyen/growmod/issues}
}

}
\author{
Jian Yen, \email{jdl.yen@gmail.com}

Freya Thomas

\strong{Maintainer}: Jian Yen \email{jdl.yen@gmail.com}

Authors:
\itemize{
  \item Freya Thomas
}

}
