% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_sim.R
\name{growmod_sim}
\alias{growmod_sim}
\title{Simulate growth curve data from a chosen growth curve}
\usage{
growmod_sim(n = 100, nblock = 5, age_range = c(0, 20),
  include_predictors = TRUE, true_model = "hillslope")
}
\arguments{
\item{n}{number of simulated growth curves}

\item{nblock}{number of blocks (e.g. species) to include in simulated data}

\item{age_range}{range of ages for simulated data}

\item{include_predictors}{logical; should growth curve parameters depend on predictor
variables that are specific to each block?}

\item{true_model}{the true growth curve model to use when simulating growth curves}
}
\value{
list containing simulated data
}
\description{
As above
}
\details{
Does XYZ
}
\examples{
\dontrun {
  sim_data <- growth_data_sim(n = 100,
                              nblock = 5,
                              age_range = c(0, 1),
                              include_predictors = TRUE,
                              true_model = 'hillslope')
}

}
