% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_methods.R
\name{plot.growmod}
\alias{plot.growmod}
\alias{plot.growmod_cv}
\alias{plot.growmod_multi}
\alias{plot.growmod_cv_multi}
\title{Plot method for \code{"growmod"} objects}
\usage{
\method{plot}{growmod}(x, y, ...)

\method{plot}{growmod_cv}(x, y, ...)

\method{plot}{growmod_multi}(x, y, group_blocks = TRUE, ...)

\method{plot}{growmod_cv_multi}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{"growmod"}, \code{"growmod_cv"},
\code{"growmod_multi"} or \code{"growmod_cv_multi"}}

\item{y}{unused default argument to plot function}

\item{\dots}{additional objects of class \code{"growmod"}, \code{"growmod_cv"},
\code{"growmod_multi"} or \code{"growmod_cv_multi"}}

\item{group_blocks}{(logical; default equals TRUE) for \code{"growmod_multi"}
or \code{"growmod_cv_multi"} objects; specify whether to plot each block
in its own plot (default) or each model in its own plot}
}
\description{
Plot method for \code{"growmod"} objects
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot fitted \code{"growmod_cv"} models

\item \code{plot}: plot fitted \code{"growmod_multi"} models

\item \code{plot}: plot fitted \code{"growmod_cv_multi"} models
}}

\examples{
\dontrun{

# simulate some data
data_test <- growmod_sim()

# fit a growmod model
mod1 <- growmod(size ~ (index | block / predictors),
                data = data_sim,
                model = 'hillslope',
                n_iter = 100,
                n_burnin = 50,
                n_chains = 1,
                stan_cores = 1)

# plot the fitted model
plot(mod1)
}

}
