% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growmod_methods.R
\name{summary.growmod}
\alias{summary.growmod}
\title{Summary method for \code{"growmod"} objects}
\usage{
\method{summary}{growmod}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"growmod"}, \code{"growmod_cv"},
\code{"growmod_multi"} or \code{"growmod_cv_multi"}}

\item{\dots}{unused additional arguments}
}
\description{
Summary method for \code{"growmod"} objects
}
\examples{
\dontrun{

# simulate some data
data_test <- growmod_sim()

# fit a growmod model
mod1 <- growmod(size ~ (index | block / predictors),
                data = data_sim,
                model = 'hillslope',
                n_iter = 100,
                n_burnin = 50,
                n_chains = 1,
                stan_cores = 1)

# summarise the fitted model
summary(mod1)
}

}
